+++ app/Exports/CustomLeagueGameImportValidation.php
@@ -5,7 +5,6 @@ namespace App\Exports;
 use App\Traits\ImportErrorHandler;
 use Maatwebsite\Excel\Concerns\Exportable;
 use Maatwebsite\Excel\Concerns\FromCollection;
-use Illuminate\Support\Facades\Log;
 use Maatwebsite\Excel\Concerns\WithStyles;
 use PhpOffice\PhpSpreadsheet\Worksheet\Worksheet;
 use PhpOffice\PhpSpreadsheet\Style\Fill;
