+++ app/Http/Controllers/UploadGameRefereesController.php
@@ -3,16 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\Imports\RefereesImport;
-use App\Imports\CustomLeagueGameImport;
 use App\Models\Region;
 use App\Traits\ImportErrorHandler;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Arr;
 use Maatwebsite\Excel\Facades\Excel;
-use Illuminate\Support\Facades\Storage;
 
-class RegionGameController extends Controller
+class UploadGameRefereesController extends Controller
 {
     use ImportErrorHandler;
 
@@ -40,7 +37,7 @@ class RegionGameController extends Controller
      * @param  \App\Models\Region  $region
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function import_referees(Request $request, $language, Region $region)
+    public function import(Request $request, $language, Region $region)
     {
         // Log::debug(print_r($request->all(),true));
         //$fname = $request->gfile->getClientOriginalName();
@@ -55,47 +52,11 @@ class RegionGameController extends Controller
             $failures = $e->failures();
             $ebag = [];
             foreach ($failures as $failure) {
-                $ebag[] = __('import.row').' "'.$failure->row().'", '.__('import.column').' "'.$failure->attribute().': '.$failure->errors()[0];
+                $ebag[] = __('import.row') . ' "' . $failure->row() . '", ' . __('import.column') . ' "' . $failure->attribute() . ': ' . $failure->errors()[0];
             }
             Log::warning('errors found in import data.', ['count' => count($failures)]);
 
             return redirect()->back()->withErrors($ebag);
-        }
-
-        return redirect()->back()->with(['status' => 'All data imported']);
-    }
-
-    /**
-     * create or update games for custom leagues
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  string  $language
-     * @param  \App\Models\Region  $region
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function import(Request $request, $language, Region $region)
-    {
-        Log::info('processing file upload.', ['region-id' => $region->id, 'file' => $request->gfile->getClientOriginalName()]);
-
-        try {
-            Log::info('validating import data.', ['region-id' => $region->id]);
-            $gImport = new CustomLeagueGameImport($region);
-            Excel::import($gImport, $request->gfile->store('temp'));
-        } catch (\Maatwebsite\Excel\Validators\ValidationException $e) {
-
-            $fileType = $request->gfile->getClientOriginalExtension();
-            if ($fileType  == 'csv') {
-                // if CSV do HTML return
-                $ebag = $this->detailedHtmlErrors(Arr::sortRecursive($e->failures()));
-                return back()->withErrors($ebag, 'default');
-            } elseif ($fileType == 'xlsx') {
-                // if excel return markedup excel file
-                $ebag = $this->excelValidationErrors($request->gfile, Arr::sortRecursive($e->failures()));
-                return back()->withErrors($ebag, 'file');
-            } else {
-                return back()->withErrors(['something went  horribly wrong :-('], 'default');
-            }
-
         }
 
         return redirect()->back()->with(['status' => 'All data imported']);
